Update week_02_Murray_poem.md
While the entire Murray reading contained valuable information, the part that stuck with me the most was the beginning when she describes the French short film called L'Arrivée d'un train en gare de La Ciotat. I love visualizing the image of an audience seeing a large train like this on the big screen for the first time. Its crazy to think that this was the foundation for immersive digital art.